+++ src/components/mod.rs
@@ -2,6 +2,11 @@ pub use self::component::*;
 pub use self::component_set::*;
 pub use self::storages::*;
 
+pub(crate) use self::grouping::*;
+pub(crate) use self::masks::*;
+
 mod component;
 mod component_set;
+mod grouping;
+mod masks;
 mod storages;