+++ src/group/mod.rs
@@ -1,8 +1,6 @@
 pub use self::info::*;
 
-pub(crate) use self::families::*;
 pub(crate) use self::mask::*;
 
-mod families;
 mod info;
 mod mask;